Add more typings + update models

+++ src/Models/Internal/Resource/Usercategoryadjustments/Patch/Patch.php
@@ -17,9 +17,9 @@ class Patch implements Model
     use ToArray;
 
     /**
-     * @var int|null Adjusted discount percentage that is applied to all products in this category.
+     * @var float|null Adjusted discount percentage that is applied to all products in this category.
      */
-    private ?int $discount;
+    private ?float $discount;
 
     /**
      * @var bool|null If TRUE then this categoy and its products won't be visible for this user.
@@ -37,9 +37,9 @@ class Patch implements Model
     }
 
     /**
-     * @return int|null Adjusted discount percentage that is applied to all products in this category.
+     * @return float|null Adjusted discount percentage that is applied to all products in this category.
      */
-    public function getDiscount(): ?int
+    public function getDiscount(): ?float
     {
         return $this->discount;
     }
@@ -61,10 +61,10 @@ class Patch implements Model
     }
 
     /**
-     * @param int|null $discount Adjusted discount percentage that is applied to all products in this category.
+     * @param float|null $discount Adjusted discount percentage that is applied to all products in this category.
      * @return self
      */
-    public function setDiscount(?int $discount): self
+    public function setDiscount(?float $discount): self
     {
         $this->discount = $discount;
         return $this;
