+++ src/Models/Vnd/Verto/Webshop/Resource/Usercategoryadjustments/Usercategoryadjustments.php
@@ -37,9 +37,9 @@ class Usercategoryadjustments implements Model
     private int $userId;
 
     /**
-     * @var int|null Adjusted discount percentage that is applied to all products in this category.
+     * @var float|null Adjusted discount percentage that is applied to all products in this category.
      */
-    private ?int $discount;
+    private ?float $discount;
 
     /**
      * @var bool If TRUE then this categoy and its products won't be visible for this user.
@@ -89,9 +89,9 @@ class Usercategoryadjustments implements Model
     }
 
     /**
-     * @return int|null Adjusted discount percentage that is applied to all products in this category.
+     * @return float|null Adjusted discount percentage that is applied to all products in this category.
      */
-    public function getDiscount(): ?int
+    public function getDiscount(): ?float
     {
         return $this->discount;
     }
@@ -153,10 +153,10 @@ class Usercategoryadjustments implements Model
     }
 
     /**
-     * @param int|null $discount Adjusted discount percentage that is applied to all products in this category.
+     * @param float|null $discount Adjusted discount percentage that is applied to all products in this category.
      * @return self
      */
-    public function setDiscount(?int $discount): self
+    public function setDiscount(?float $discount): self
     {
         $this->discount = $discount;
         return $this;
