+++ src/Models/Vnd/Verto/Webshop/Resource/Usergroupcategoryadjustments/Usergroupcategoryadjustments.php
@@ -37,9 +37,9 @@ class Usergroupcategoryadjustments implements Model
     private int $usergroupId;
 
     /**
-     * @var int|null Adjusted discount percentage that is applied to all products in this category.
+     * @var float|null Adjusted discount percentage that is applied to all products in this category.
      */
-    private ?int $discount;
+    private ?float $discount;
 
     /**
      * @var bool If TRUE then this categoy and its products won't be visible for this user group.
@@ -89,9 +89,9 @@ class Usergroupcategoryadjustments implements Model
     }
 
     /**
-     * @return int|null Adjusted discount percentage that is applied to all products in this category.
+     * @return float|null Adjusted discount percentage that is applied to all products in this category.
      */
-    public function getDiscount(): ?int
+    public function getDiscount(): ?float
     {
         return $this->discount;
     }
@@ -153,10 +153,10 @@ class Usergroupcategoryadjustments implements Model
     }
 
     /**
-     * @param int|null $discount Adjusted discount percentage that is applied to all products in this category.
+     * @param float|null $discount Adjusted discount percentage that is applied to all products in this category.
      * @return self
      */
-    public function setDiscount(?int $discount): self
+    public function setDiscount(?float $discount): self
     {
         $this->discount = $discount;
         return $this;
